+++ ui/debugger/panel.go
@@ -1,7 +1,7 @@
 package debugger
 
 import (
-	"github.com/danielecanzoneri/gb-emulator/ui/theme"
+	"github.com/danielecanzoneri/gb-emulator/ui/graphics"
 	"image/color"
 
 	"github.com/danielecanzoneri/gb-emulator/gameboy"
